Add files via upload
1.Simple Brilliant Mind Game Jackpot Program:
Using nested loops to build a Java program that handles a game called "Brilliant Mind". The game involves processing scores from different categories like Science, Literature, Sports, etc., each with a unique scoring system. The scores for each category are computed based on numbers that are not multiples of the category number. The results are displayed, and the score is cumulatively updated as the player progresses through the categories.
2.Simple Exam Registration Program:
This task involves using arrays to manage a registration system for exams, specifically tailored for the Access Center for Students with Disabilities at Concordia. The program should allow for listing courses, adding students and courses, and displaying course details. The user interactions are handled through a menu-driven interface, where inputs are validated, and the list of courses can be updated and displayed based on the user's choices.